trying a pausable stopwatch with fixed capture rate

+++ Object Physics Sandbox/Assets/Save.cs
@@ -8,9 +8,9 @@ public class Save
     public List<Vector3> velocities = new List<Vector3>();
     public List<Vector3> ballPositions = new List<Vector3>();
     public List<Vector3> objectPositions = new List<Vector3>();
-    public List<long> velocitiesCT = new List<long>();
-    public List<long> ballPositionsCT = new List<long>();
-    public List<long> objectPositionsCT = new List<long>();
+    public List<double> velocitiesCT = new List<double>();
+    public List<double> ballPositionsCT = new List<double>();
+    public List<double> objectPositionsCT = new List<double>();
     public List<string> foundObjectsTags = new List<string>();
 
 }
